initial ts refactor
