+++ .babelrc
@@ -1,1 +1,1 @@
-{ "presets": ["@babel/preset-env"] }
+{ "presets": ["@babel/preset-env", "@babel/preset-typescript"] }
