+++ package.json
@@ -27,8 +27,11 @@
     "postpublish": "run-s gh-pages",
     "gh-pages": "gh-pages -d pages",
     "copyright": "update license",
-    "lint": "eslint src --ext .mjs",
-    "build": "rollup -c",
+    "lint": "eslint src --ext .ts",
+    "check": "tsc --noEmit -p tsconfig.json",
+    "build:types": "tsc -p tsconfig.types.json",
+    "build:js": "rollup -c",
+    "build": "run-s build:js build:types",
     "build:min": "rollup -c",
     "build:release": "run-s build build:min copyright",
     "watch": "rollup -c --watch",
@@ -51,21 +54,26 @@
   "devDependencies": {
     "@babel/core": "^7.29.0",
     "@babel/preset-env": "^7.29.2",
+    "@babel/preset-typescript": "^7.27.0",
     "@dagrejs/dagre": "^3.0.0",
     "@rollup/plugin-babel": "^7.0.0",
     "@rollup/plugin-commonjs": "^29.0.2",
     "@rollup/plugin-node-resolve": "^16.0.3",
     "@rollup/plugin-terser": "^1.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
     "camelcase": "^6.0.0",
     "chai": "4.2.0",
     "cpy-cli": "^3.1.1",
     "cross-env": "^7.0.2",
+    "cytoscape": "^3.2.22",
     "eslint": "^7.6.0",
     "gh-pages": "^5.0.0",
     "mocha": "8.1.1",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
     "rollup": "^4.60.4",
+    "typescript": "^5.4.0",
     "update": "^0.7.4",
     "updater-license": "^1.0.0"
   },
